Added status code

+++ Controller/controller.js
@@ -3,9 +3,9 @@ const db = require("../config/db");
 const getUser = (req, res) => {
   db.query("select * from USERS1", function (err, result, fields) {
     if (err) {
-      console.log(err);
+      throw err;
     } else {
-      res.send(result);
+      res.status(200).json(result);
     }
   });
 };
@@ -19,13 +19,13 @@ const createUser = (req, res) => {
   const img = req.body.url;
   db.query(
     `INSERT INTO USERS1(NAME,LAST_NAME,EMAIL,USER_ID,MOBILE,IMG_URL) VALUES ("${name}","${lastname}","${email}","${id}","${phone}","${img}")`,
-    (err) => {
+    (err, result, fields) => {
       if (err) {
-        console.log(err);
         res.send("ERROR IN INSERT!!!!");
       } else {
-        console.log("INSERT SUCCESSFUL");
-        res.send("INSERTED!!!!");
+        return res.status(200).json({
+          message: "User created!!",
+        });
       }
     }
   );
@@ -40,13 +40,16 @@ const updateUser = (req, res) => {
   console.log(id);
   db.query(
     `UPDATE USERS1 SET NAME = "${name}", LAST_NAME = "${lastname}", EMAIL = "${email}", MOBILE = "${phone}" WHERE USER_ID="${id}"`,
-    (err) => {
+    (err, result) => {
       if (err) {
         console.log(err);
         res.send("ERROR IN INSERT!!!!");
       } else {
-        console.log("INSERT SUCCESSFUL");
-        res.send("INSERTED!!!!");
+        return res.status(200).json({
+          error: false,
+          data: result,
+          message: "users list.",
+        });
       }
     }
   );
@@ -54,11 +57,21 @@ const updateUser = (req, res) => {
 
 const deleteUser = (req, res) => {
   const delId = req.params.id;
+  if (!delId) {
+    return res.status(400).json({
+      error: true,
+      message: "Please provide userId",
+    });
+  }
   db.query("delete from USERS1 where USER_ID=?", [delId], (err, result) => {
     if (err) {
       console.log(err);
     } else {
-      res.send("deleted");
+      console.log(result);
+      return res.status(200).json({
+        status: "success",
+        message: "User is deleted",
+      });
     }
   });
 };
